% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Illustration of the Euclidean algorithm that is finding the greatest common denominator between two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Numeric value that is bigger than b}

\item{b}{Numeric value that is smaller than a}
}
\value{
greatest common divisor of numbers a and b is found
}
\description{
Illustration of the Euclidean algorithm that is finding the greatest common denominator between two numbers
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
